requirements.txt
python-dotenv
flask==3.0.3
PyMuPDF==1.24.9
openai
pdf2image
pytesseract
pillow
requests